Converting to Postgresql - still in progress. Requiring changes across a broad number of functions/files
Note:
 - Currently database is only holding Disease, Feat, Item and Power info.  Will implement search and bot functions before completing the rest of the parsing.

TODO:
Switch to Postgresql from SQLite
Initiative Tracker
Improved Bot Search
Parers for rest of data
Scrape HTML data

+++ database_models.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
-from sqlalchemy import Integer
+from sqlalchemy import Integer, BigInteger
 from sqlalchemy import String, Boolean
 from sqlalchemy.orm import declarative_base
 from sqlalchemy.orm import relationship
@@ -13,8 +13,8 @@ Base = declarative_base()
 class Global(Base):
     __tablename__ = "global_manager"
 
-    guild_id = Column(Integer(), primary_key=True, autoincrement=False)
-    time = Column(Integer(), default=0, nullable=False)
+    guild_id = Column(BigInteger(), primary_key=True, autoincrement=False)
+    time = Column(BigInteger(), default=0, nullable=False)
     initiative = Column(Integer())
     saved_order = Column(String())
     gm = Column(String())
